app/requirements.txt
Django==4.2
django-extensions==3.2.1
Pillow==9.5.0
django-cas-ng==4.3.0
psycopg==3.1.9
python-dotenv==1.0.0
djangorestframework==3.14.0
drf-yasg==1.21.7